refactor to use __init__ and __main__ a bit more correctly
